+++ client/src/app/components/rdio-scanner/rdio-scanner.ts
@@ -49,7 +49,6 @@ export interface RdioScannerCall {
     frequencies?: RdioScannerCallFrequency[];
     frequency?: number;
     id: number;
-    patched?: boolean;
     patches: number[];
     source?: number;
     sources?: RdioScannerCallSource[];